Create shared onIncompatibleSchemaChange Autocomplete component
